+++ app/components/primer/details_component.rb
@@ -19,11 +19,11 @@ module Primer
     with_slot :body, class_name: "Body"
     with_slot :summary, class_name: "Summary"
 
-    def initialize(overlay: NO_OVERLAY, reset: false, **kwargs)
-      @kwargs = kwargs
-      @kwargs[:tag] = :details
-      @kwargs[:classes] = class_names(
-        kwargs[:classes],
+    def initialize(overlay: NO_OVERLAY, reset: false, **system_arguments)
+      @system_arguments = system_arguments
+      @system_arguments[:tag] = :details
+      @system_arguments[:classes] = class_names(
+        system_arguments[:classes],
         OVERLAY_MAPPINGS[fetch_or_fallback(OVERLAY_MAPPINGS.keys, overlay, NO_OVERLAY)],
         "details-reset" => reset
       )
@@ -34,28 +34,28 @@ module Primer
     end
 
     class Summary < Primer::Slot
-      def initialize(button: true, **kwargs)
+      def initialize(button: true, **system_arguments)
         @button = button
 
-        @kwargs = kwargs
-        @kwargs[:tag] = :summary
-        @kwargs[:role] = "button"
+        @system_arguments = system_arguments
+        @system_arguments[:tag] = :summary
+        @system_arguments[:role] = "button"
       end
 
       def component
-        return Primer::BaseComponent.new(**@kwargs) unless @button
+        return Primer::BaseComponent.new(**@system_arguments) unless @button
 
-        Primer::ButtonComponent.new(**@kwargs)
+        Primer::ButtonComponent.new(**@system_arguments)
       end
     end
 
     class Body < Primer::Slot
       DEFAULT_OMIT_WRAPPER = false
 
-      def initialize(omit_wrapper: DEFAULT_OMIT_WRAPPER, **kwargs)
+      def initialize(omit_wrapper: DEFAULT_OMIT_WRAPPER, **system_arguments)
         @omit_wrapper = fetch_or_fallback([true, false], omit_wrapper, DEFAULT_OMIT_WRAPPER)
-        @kwargs = kwargs
-        @kwargs[:tag] ||= :div
+        @system_arguments = system_arguments
+        @system_arguments[:tag] ||= :div
       end
 
       def omit_wrapper?
@@ -63,7 +63,7 @@ module Primer
       end
 
       def component
-        Primer::BaseComponent.new(**@kwargs)
+        Primer::BaseComponent.new(**@system_arguments)
       end
     end
   end