+++ app/components/primer/flex_item_component.rb
@@ -5,17 +5,17 @@ module Primer
     FLEX_AUTO_DEFAULT = false
     FLEX_AUTO_ALLOWED_VALUES = [FLEX_AUTO_DEFAULT, true]
 
-    def initialize(flex_auto: FLEX_AUTO_DEFAULT, **kwargs)
-      @kwargs = kwargs
-      @kwargs[:classes] =
+    def initialize(flex_auto: FLEX_AUTO_DEFAULT, **system_arguments)
+      @system_arguments = system_arguments
+      @system_arguments[:classes] =
         class_names(
-          @kwargs[:classes],
+          @system_arguments[:classes],
           "flex-auto" => fetch_or_fallback(FLEX_AUTO_ALLOWED_VALUES, flex_auto, FLEX_AUTO_DEFAULT)
         )
     end
 
     def call
-      render(Primer::BoxComponent.new(**@kwargs)) { content }
+      render(Primer::BoxComponent.new(**@system_arguments)) { content }
     end
   end
 end