+++ app/components/primer/label_component.rb
@@ -3,12 +3,18 @@
 module Primer
   # Use labels to add contextual metadata to a design.
   class LabelComponent < Primer::Component
-    SCHEME_MAPPINGS = {
-      # gray
+    NEW_SCHEME_MAPPINGS = {
+      primary: "Label--primary",
+      secondary: "Label--secondary",
+      info: "Label--info",
+      success: "Label--success",
+      warning: "Label--warning",
+      danger: "Label--danger",
+    }
+
+    DEPRECATED_SCHEME_MAPPINGS = {
       gray: "Label--gray",
       dark_gray: "Label--gray-darker",
-
-      # colored
       yellow: "Label--yellow",
       orange: "Label--orange",
       red: "Label--red",
@@ -16,11 +22,11 @@ module Primer
       blue: "Label--blue",
       purple: "Label--purple",
       pink: "Label--pink",
-
-      # Deprecated
       outline: "Label--outline",
       green_outline: "Label--outline-green",
     }.freeze
+
+    SCHEME_MAPPINGS = NEW_SCHEME_MAPPINGS.merge(DEPRECATED_SCHEME_MAPPINGS)
     SCHEME_OPTIONS = SCHEME_MAPPINGS.keys << nil
 
     VARIANT_MAPPINGS = {
@@ -42,24 +48,24 @@ module Primer
     #   <%= render(Primer::LabelComponent.new(title: "Label: Label", variant: :large)) { "Large" } %>
     #
     # @param title [String] `title` attribute for the component element.
-    # @param scheme [Symbol] <%= one_of(Primer::LabelComponent::SCHEME_OPTIONS) %>
+    # @param scheme [Symbol] <%= one_of(Primer::LabelComponent::DEPRECATED_SCHEME_MAPPINGS.keys) %>
     # @param variant [Symbol] <%= one_of(Primer::LabelComponent::VARIANT_OPTIONS) %>
-    # @param kwargs [Hash] <%= link_to_style_arguments_docs %>
-    def initialize(title:, scheme: nil, variant: nil, **kwargs)
-      @kwargs = kwargs
-      @kwargs[:bg] = :blue if scheme.nil?
-      @kwargs[:tag] ||= :span
-      @kwargs[:title] = title
-      @kwargs[:classes] = class_names(
+    # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
+    def initialize(title:, scheme: nil, variant: nil, **system_arguments)
+      @system_arguments = system_arguments
+      @system_arguments[:bg] = :blue if scheme.nil?
+      @system_arguments[:tag] ||= :span
+      @system_arguments[:title] = title
+      @system_arguments[:classes] = class_names(
         "Label",
-        kwargs[:classes],
+        system_arguments[:classes],
         SCHEME_MAPPINGS[fetch_or_fallback(SCHEME_OPTIONS, scheme)],
         VARIANT_MAPPINGS[fetch_or_fallback(VARIANT_OPTIONS, variant)]
       )
     end
 
     def call
-      render(Primer::BaseComponent.new(**@kwargs)) { content }
+      render(Primer::BaseComponent.new(**@system_arguments)) { content }
     end
   end
 end