+++ app/components/primer/octicon_component.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Primer
-  # Renders an [Octicon](https://primer.style/octicons/) with <%= link_to_style_arguments_docs %>.
+  # Renders an [Octicon](https://primer.style/octicons/) with <%= link_to_system_arguments_docs %>.
   class OcticonComponent < Primer::Component
     include Primer::ClassNameHelper
     include OcticonsHelper
@@ -24,22 +24,22 @@ module Primer
     #   <%= render(Primer::OcticonComponent.new(icon: "x", size: :large)) %>
     #
     # @param icon [String] Name of [Octicon](https://primer.style/octicons/) to use.
-    # @param size [Symbol] <%= one_of(Primer::OcticonComponent::SIZE_OPTIONS) %>
-    # @param kwargs [Hash] <%= link_to_style_arguments_docs %>
-    def initialize(icon:, size: SIZE_DEFAULT, **kwargs)
-      @icon, @kwargs = icon, kwargs
+    # @param size [Symbol] <%= one_of(Primer::OcticonComponent::SIZE_MAPPINGS) %>
+    # @param system_arguments [Hash] <%= link_to_system_arguments_docs %>
+    def initialize(icon:, size: SIZE_DEFAULT, **system_arguments)
+      @icon, @system_arguments = icon, system_arguments
 
-      @kwargs[:class] = Primer::Classify.call(**@kwargs)[:class]
-      @kwargs[:height] ||= SIZE_MAPPINGS[size]
+      @system_arguments[:class] = Primer::Classify.call(**@system_arguments)[:class]
+      @system_arguments[:height] ||= SIZE_MAPPINGS[size]
 
       # Filter out classify options to prevent them from becoming invalid html attributes.
       # Note height and width are both classify options and valid html attributes.
-      octicon_helper_options = @kwargs.slice(:height, :width)
-      @kwargs = @kwargs.except(*Primer::Classify::VALID_KEYS, :classes).merge(octicon_helper_options)
+      octicon_helper_options = @system_arguments.slice(:height, :width)
+      @system_arguments = @system_arguments.except(*Primer::Classify::VALID_KEYS, :classes).merge(octicon_helper_options)
     end
 
     def call
-      octicon(@icon, **@kwargs)
+      octicon(@icon, **@system_arguments)
     end
   end
 end